chore: update styles
