+++ src/components/cards/categoryCard/CategoryCard.module.scss
@@ -1,0 +1,14 @@
+.box {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+  text-align: center;
+  border: 1px solid rgba(0, 0, 0, 0.3);
+  border-radius: 4px;
+}
+
+.box:hover {
+  cursor: pointer;
+}
