+++ src/components/cards/categoryCard/assets/droid.png
