+++ src/components/cards/categoryCard/assets/farming.png
