+++ src/components/cards/categoryCard/assets/kitchen.png
