+++ src/components/cards/categoryCard/assets/vacuum.png
