+++ src/components/cards/productCard/ProductCard.module.scss
@@ -1,0 +1,5 @@
+.wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
